player: Move coil-, switch-, lamp-, and wire handling into separate classes.
