+++ VisualPinball.Unity/VisualPinball.Unity/Game/CoilPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9d3730a4e826534c855eee8061e99d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
