+++ VisualPinball.Unity/VisualPinball.Unity/Game/LampPlayer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cf8ea31819f4e2b47b6a5ccd3cfce295
+guid: c31c8b5b6e7f19e4880f9148f78e1c4b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
