+++ VisualPinball.Unity/VisualPinball.Unity/Game/WirePlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47ef89abae693fe43acad81109183617
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
